checkpoint

+++ src/resources/item.py
@@ -1,8 +1,8 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
-from models.item import ItemModel
+from src.models.item import ItemModel
 
 
 class Item(Resource):
@@ -50,7 +50,13 @@ class Item(Resource):
         # 201 is for created
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required"}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
